feat: collapsible bottom panel
