+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Collapsible `Send/Console` panel
+
+### Changed
+
+- Save layout state when closing application
+
 ## [0.9.6] - 2025-09-25
 
 ### Added
